refactor: application-list 폴더 구종 정리 (#66)
