+++ apps/web/src/app/(main)/application-list/setting/(preview)/PreviewComponent.css.ts
