+++ apps/web/src/app/(main)/application-list/setting/(preview)/_components/ApplicationPartPreview/ApplicationPartsPreview.tsx
