+++ apps/web/src/app/(main)/application-list/setting/(preview)/_components/BasicInfoPreview/BasicInfoPreview.css.ts
