+++ apps/web/src/app/(main)/application-list/setting/[id]/preview/page.tsx
@@ -1,0 +1,5 @@
+import PreviewComponent from '../../(preview)/PreviewComponent';
+
+export default function EditPreviewPage() {
+  return <PreviewComponent />;
+}
