+++ apps/web/src/app/(main)/application-list/setting/new/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-import React from 'react';
 import { SettingForm } from '@web/app/(main)/application-list/setting/_components/SettingForm/SettingForm';
 
 export default function NewSettingPage() {
