+++ apps/web/src/app/(main)/application-list/setting/new/preview/page.tsx
@@ -1,0 +1,6 @@
+import PreviewComponent from "../../(preview)/PreviewComponent";
+
+export default function NewPreviewPage() {
+    return <PreviewComponent />
+  }
+
