+++ apps/web/src/app/(main)/apply-management/add/_components/BasicInfoForm/BasicInfoForm.tsx
@@ -5,8 +5,8 @@ import { parseISO } from 'date-fns';
 import { DatePicker } from '@repo/ui/DatePicker';
 import { IcImage } from '@repo/ui/icons/colored';
 import { Option } from '@repo/ui/Option';
-import { InfoField } from '@web/app/(main)/application-list/setting/preview/_components/InfoField/InfoField';
-import { DateChip } from '@web/app/(main)/application-list/setting/preview/_components/DateChip/DateChip';
+import { InfoField } from '@web/components/InfoField/InfoField';
+import { DateChip } from '@web/components/DateChip/DateChip';
 import { TextField } from '@repo/ui/InputField';
 import Image from 'next/image';
 import * as styles from './BasicInfoForm.css';
